сверстан экран настроек
