+++ TravelSchedule/Helpers/NavigationConstants.swift
@@ -1,0 +1,12 @@
+//
+//  NavigationConstants.swift
+//  TravelSchedule
+//
+//  Created by Olga Trofimova on 23.02.2025.
+//
+
+import Foundation
+
+enum NavigationConstants: String {
+    case userAgreementView = "UserAgreementView"
+}
