+++ TravelSchedule/Views/Settings/SettingsView.swift
@@ -8,11 +8,42 @@
 import SwiftUI
 
 struct SettingsView: View {
+    
+    @State private var isDarkModeOn = false
+    @Binding var path: [String]
+    
     var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+            VStack {
+                Toggle("Темная тема", isOn: $isDarkModeOn)
+                    .toggleStyle(SwitchToggleStyle(tint: .ypBlue))
+                    .frame(height: 60)
+                
+                    HStack() {
+                        Text("Пользовательское соглашение")
+                        Spacer()
+                        Image("Chevron")
+                    }
+                    .background()
+                    .frame(height: 60)
+                    .onTapGesture {
+                        print("tap")
+                        path.append("UserAgreementView")
+                    }
+                
+                Spacer()
+                
+                VStack(spacing: 12) {
+                    Text("Приложение использует API «Яндекс.Расписания»")
+                    Text("Версия 1.0 (beta)")
+                }
+                .foregroundStyle(.ypBlack)
+                .font(.system(size: 12, weight: .regular))
+            }
+            .foregroundStyle(.ypBlack)
+            .padding(16)
     }
 }
 
 #Preview {
-    SettingsView()
+    SettingsView(path: .constant([""]))
 }
