+++ TravelSchedule/Views/Settings/UserAgreementView.swift
@@ -10,6 +10,9 @@ import SwiftUI
 struct UserAgreementView: View {
     var body: some View {
         Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+            .navigationTitle("Пользовательское соглашение")
+            .navigationBarTitleDisplayMode(.inline)
+            .tint(.ypBlack)
     }
 }
